// kitabaca/src/interfaces/userInterface.d.ts
export interface IUser{
     UserID? : number;
     FullName? : string;
     DOB? : string;
     Email? : string;
     Password? : string;
     ConfirmPassword? : string;
     ProfilePict? : string;
     TotalCoin? : number;
     Level? : number;
     Role? : string;
     SaveBookID? : number;
     SavePage? : number;
     BookCategory? : string;
}